feat: Provide active CSS class to the current page
This replaces the old method of checking based on the page name, which
generally increased the size of all pages (and the CSS) by a couple
dozen bytes and is less flexible. Nav pages are also determined
programmatically from the data file.

Autoformatter hit the default layout, sorry about that.